refactor(test): move test utils to utils/testing
