+++ repo/schema/schema_manager.go
@@ -1,4 +1,4 @@
-package util
+package schema
 
 import (
 	"fmt"
