+++ repo/schema/schema_manager_test.go
@@ -1,4 +1,4 @@
-package util
+package schema
 
 import (
 	"os"
